requirements.txt
python-telegram-bot[ext]==20.7
openai==1.82.0
psycopg2-binary==2.9.9
aiohttp==3.9.5
python-dotenv==1.0.1
APScheduler==3.10.4
